pyproject.toml
[tool.poetry]
name = "dict-to-yaml"
version = "0.1.0"
description = ""
authors = ["cetinsamet <cet.samet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.33.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
